+++ packages/runtime/src/types.ts
@@ -88,6 +88,15 @@ export interface DataLoaderConfig {
   options?: DataLoaderOptions;
 }
 
+interface DocumentLoaderOptions {
+  documentOnly?: boolean;
+}
+export type DocumentDataLoader = (ctx: RequestContext, options: DocumentLoaderOptions) => DataLoaderResult;
+
+export interface DocumentDataLoaderConfig {
+  loader: DocumentDataLoader;
+}
+
 export interface LoadersData {
   [routeId: string]: LoaderData;
 }
@@ -312,7 +321,7 @@ export interface ServerRenderOptions {
   assetsManifest: AssetsManifest;
   createRoutes: (options: Pick<RouteLoaderOptions, 'requestContext' | 'renderMode'>) => RouteItem[];
   runtimeModules: RuntimeModules;
-  documentDataLoader?: DataLoaderConfig;
+  documentDataLoader?: DocumentDataLoaderConfig;
   Document: DocumentComponent;
   documentOnly?: boolean;
   renderMode?: RenderMode;